修复单线程断点续传异常

+++ model_multiload/src/main/java/com/wukj/general/multiload/RateListener.java
@@ -17,5 +17,5 @@ package com.wukj.general.multiload;
  * 3.
  */
 public interface RateListener {
-    void onDownloadSize(int progress);
+    void onDownloadSize(float progress);
 }
